Refactor codex tab into dedicated module
